* commit 'da656a0b968f0fc2c61b71af5805270e2874fde6':
  - All warnings in the RobotWorld's own code (but one) are gone - The remaining warning is in NotificationHandler.hpp, line 22/23. I don't have a solution yet. - Client.hpp and Server.hpp have embeded PlantUML. See http://plantuml.com/eclipse for an PlantUML Eclipse plugin.

+++ src/AStar.cpp
@@ -118,7 +118,7 @@ namespace PathAlgorithm
 		getCS().clear();
 		getPM().clear();
 
-		int radius = std::ceil(std::sqrt( (aRobotSize.x / 2.0) * (aRobotSize.x / 2.0) + (aRobotSize.y / 2.0) * (aRobotSize.y / 2.0)));
+		int radius = static_cast<int>(std::ceil(std::sqrt( (aRobotSize.x / 2.0) * (aRobotSize.x / 2.0) + (aRobotSize.y / 2.0) * (aRobotSize.y / 2.0))));
 
 		aStart.actualCost = 0.0; 													// Cost from aStart along the best known path.
 		aStart.heuristicCost = aStart.actualCost + HeuristicCost( aStart, aGoal);	// Estimated total cost from aStart to aGoal through y.
